requirements-dev.txt
fastapi==0.122.0
uvicorn==0.38.0
requests==2.32.5
python-dotenv==1.0.0
cryptography==41.0.7
pyjwt==2.10.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
